requirements.txt
braceexpand==0.1.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
distlib==0.3.6
faiss-gpu==1.7.2
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
ftfy==6.1.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Levenshtein==0.20.9
MarkupSafe==2.1.2
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
open-clip-torch==0.2.1
Pillow==9.4.0
pipenv==2022.12.19
platformdirs==2.6.2
PyYAML==6.0
rapidfuzz==2.13.7
regex==2022.10.31
requests==2.28.2
six==1.16.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
virtualenv==20.17.1
virtualenv-clone==0.5.7
wcwidth==0.2.6
webdataset==0.2.33
Werkzeug==2.2.2
zipp==3.13.0
